Use VecMap instead of HashMap to store variables
This is even faster than HashMap, and since variable names are unsigned numbers this is a perfect use case for VecMap

+++ src/literal.rs
@@ -21,7 +21,7 @@ impl Not for Literal {
 
 impl Literal {
     pub fn new(num: &i64) -> Literal {
-        Literal(num.abs() as u64, *num >= 0)
+        Literal(num.abs() as usize, *num >= 0)
     }
 
     pub fn falsified_by(&self, var_val: VariableState) -> bool {
